refactoring DataReport generator code and adding AlertV2 to existing integ test suit.
