+++ thirdeye/thirdeye-pinot/src/test/java/com/linkedin/thirdeye/datalayer/bao/AbstractManagerTestBase.java
@@ -15,7 +15,9 @@ import com.linkedin.thirdeye.datalayer.bao.jdbc.MetricConfigManagerImpl;
 import com.linkedin.thirdeye.datalayer.bao.jdbc.OverrideConfigManagerImpl;
 import com.linkedin.thirdeye.datalayer.bao.jdbc.RawAnomalyResultManagerImpl;
 import com.linkedin.thirdeye.datalayer.bao.jdbc.TaskManagerImpl;
+import com.linkedin.thirdeye.datalayer.dto.AlertConfigDTO;
 import com.linkedin.thirdeye.datalayer.dto.OverrideConfigDTO;
+import com.linkedin.thirdeye.datalayer.pojo.AlertConfigBean;
 import com.linkedin.thirdeye.datalayer.util.DaoProviderUtil;
 
 import com.linkedin.thirdeye.detector.metric.transfer.ScalingFactor;
@@ -173,6 +175,23 @@ public abstract class AbstractManagerTestBase {
     functionSpec.setProperties("baseline=w/w;changeThreshold=0.001");
     functionSpec.setIsActive(true);
     return functionSpec;
+  }
+
+  protected AlertConfigDTO getTestAlertConfiguration(String name) {
+    AlertConfigDTO alertConfigDTO = new AlertConfigDTO();
+    alertConfigDTO.setName(name);
+    alertConfigDTO.setActive(true);
+    alertConfigDTO.setFromAddress("te@linkedin.com");
+    alertConfigDTO.setRecipients("anomaly@linedin.com");
+    alertConfigDTO.setCronExpression("0/10 * * * * ?");
+    AlertConfigBean.EmailConfig emailConfig = new AlertConfigBean.EmailConfig();
+    emailConfig.setLastNotifiedAnomalyId(0l);
+    emailConfig.setSendAlertOnZeroAnomaly(true);
+    alertConfigDTO.setEmailConfig(emailConfig);
+    AlertConfigBean.ReportConfig reportConfig = new AlertConfigBean.ReportConfig();
+    reportConfig.setEnabled(true);
+    alertConfigDTO.setReportConfig(reportConfig);
+    return alertConfigDTO;
   }
 
   protected EmailConfigurationDTO getTestEmailConfiguration(String metricName, String collection) {
